+++ tests/test_freeze_dynamic.py
@@ -158,7 +158,7 @@ def test_freeze_bytes_roundtrip():
 def test_freeze_load_sample(tmpdir):
     o = tmpdir.mkdir("capa").join("test.frz")
 
-    extractor = fixtures.get_cape_extractor(fixtures.get_data_path_by_name("d46900"))
+    extractor = fixtures.get_cape_extractor(fixtures.get_data_path_by_name("d4690..."))
 
     Path(o.strpath).write_bytes(capa.features.freeze.dump(extractor))